+++ mettagrid/src/metta/mettagrid/stats_tracker.hpp
@@ -119,7 +119,7 @@ public:
       auto first_it = _first_seen_at.find(key);
       auto last_it = _last_seen_at.find(key);
       if (first_it != _first_seen_at.end() && last_it != _last_seen_at.end()) {
-        int duration = last_it->second - first_it->second;
+        int duration = static_cast<int>(last_it->second) - static_cast<int>(first_it->second);
         if (duration > 0 && count > 1) {
           result[key + ".activity_rate"] = static_cast<float>(count - 1) / static_cast<float>(duration);
         }
